move index to home, Index to HomeContent
