+++ paywall/src/__tests__/components/content/HomeContent.test.js
@@ -2,20 +2,20 @@ import React from 'react'
 import * as rtl from 'react-testing-library'
 import { Provider } from 'react-redux'
 
-import Index from '../../components/Index'
-import createUnlockStore from '../../createUnlockStore'
-import { ConfigContext } from '../../utils/withConfig'
-import { WindowContext } from '../../hooks/browser/useWindow'
+import HomeContent from '../../../components/content/HomeContent'
+import createUnlockStore from '../../../createUnlockStore'
+import { ConfigContext } from '../../../utils/withConfig'
+import { WindowContext } from '../../../hooks/browser/useWindow'
 
 jest.useFakeTimers()
-describe('Index component', () => {
+describe('Paywall home page component', () => {
   const lock = '0x5Cd3FC283c42B4d5083dbA4a6bE5ac58fC0f0267'
   const account = '0xAaAdEED4c0B861cB36f4cE006a9C90BA2E43fdc2'
   const transaction =
     '0x1bd7cf2e0d9f7ede3473ba1588fa5258a4920d57e9c97cd10350c7482d3cbff2'
   it('should return the landing page if no lock is passed', () => {
     expect.assertions(2)
-    const component = rtl.render(<Index path="/" />)
+    const component = rtl.render(<HomeContent path="/" />)
 
     expect(component.getByText('Pay for Content Seamlessly')).not.toBeNull()
     expect(component.queryByText('30 days')).toBeNull()
@@ -69,7 +69,7 @@ describe('Index component', () => {
       <Provider store={store}>
         <WindowContext.Provider value={window}>
           <ConfigContext.Provider value={config}>
-            <Index path={`/${lock}`} />
+            <HomeContent path={`/${lock}`} />
           </ConfigContext.Provider>
         </WindowContext.Provider>
       </Provider>
